+++ components/displaypage/DisplayPageLayout.js
@@ -1,5 +1,6 @@
 import React from "react";
 import NavBar from "../NavBar/NavBar";
+import styles from "./DisplayPageLayout.module.css";
 
 export default function DisplayPageLayout ({children}) {
 
@@ -7,7 +8,7 @@ export default function DisplayPageLayout ({children}) {
     <div className="container">
         <div><NavBar /></div>
         <div className="d-flex">
-            <div>SideBar</div>
+            <div className={styles.sideBar}>SideBar</div>
             <div style={{flex:1}}>{children}</div>
             
         </div>
