+++ components/displaypage/DisplayPageLayout.module.css
@@ -1,0 +1,4 @@
+.sideBar {
+    display: none;
+
+  }
