+++ components/displaypage/DisplayPost.module.css
@@ -2,11 +2,11 @@
     display: grid;
     width: 100%;
     grid-template-columns: 1fr !important; 
-
     grid-gap:10px;
     grid-template-areas:
-        "content"
+        "title"
         "photoGallery"
+        "content"
         "map"
 }
 
@@ -16,13 +16,16 @@
 }
 
 .content {
-    grid-area: content
+    grid-area: content;
 }
 
 .map {
-    grid-area: map
+    grid-area: map;
 }
 
+.title {
+    grid-area: title;
+}
 
 /*  landscape mobile */
 @media (min-width:768px) {
@@ -31,7 +34,14 @@
 
 /* tablet, */
 @media (min-width:1024px) {
-    
+    .mainField {
+        grid-template-columns: 2fr 1fr !important; 
+        grid-gap:10px;
+        grid-template-areas:
+            "title"
+            "photoGallery content"
+            "map content"
+    }
 }
 
 /* desktop */
@@ -40,6 +50,7 @@
         grid-template-columns: 2fr 1fr !important; 
         grid-gap:10px;
         grid-template-areas:
+            "title"
             "photoGallery content"
             "map content"
     }
