+++ components/displaypage/Map.module.css
@@ -1,4 +1,4 @@
 .leaflet-container {
-    width: 50%;
-    height: 70vh;
+    width: 100%;
+    height: 100vh;
   }
